pyproject.toml
[project]
name = "sovereign-rag-gateway"
version = "0.1.0"
description = "Policy-first OpenAI-compatible gateway for regulated workloads"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "jsonschema>=4.23.0,<5.0.0",
  "httpx>=0.27.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "mypy>=1.11.0,<2.0.0",
  "ruff>=0.6.0,<1.0.0",
  "openai>=1.52.0,<2.0.0",
  "types-jsonschema>=4.23.0.20241208",
  "types-requests>=2.32.0.20241016"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict_equality = true
packages = ["app", "scripts"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
